move DragDropContext to FutureView index
